cli: move addrSetter into package `addr`
This makes it available for use by other go programs that
don't want to depend on `cli`, such as `cockroach-sql`.

Release note: None
